+++ packages/react/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Nimbus is an open-source Design System created by Tiendanube / Nuvemshop's team to empower and enhance more stories every day, with simplicity, accessibility, consistency and performance.
 
+## 2025-11-13 `5.37.0`
+
+#### 🎉 New features
+
+- Added `large` size to `Button` component. ([#380](https://github.com/TiendaNube/nimbus-design-system/pull/380) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2025-11-11 `5.36.0`
 
 #### 🎉 New features